- fixed memory[total] for HP-UX when >4GB of memory (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@1710 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/sysinfo.c
@@ -1802,7 +1802,7 @@ int	TOTALMEM(const char *cmd, const char *parameter,double  *value)
 		/* Get page size */	
 		page = pst.page_size;
 		/* Total physical memory in bytes */	
-		*value=page*pst.physical_memory;
+		*value=(double)page*pst.physical_memory;
 		return SYSINFO_RET_OK;
 	}
 #else
